// UniVe.TAW.Mobile/src/assets/unive.taw.webservice/application/DTOs.ts
export interface IUserDto {
    Id: string;
    Username: string;
    Age: number;
    CountryId: enums.Country;
    BannedUntil: Date;
    Roles: enums.UserRoles;
}
// ...
export interface IMatchSnapshotDto {
    Id: string;
    Settings: game.IMatchSettings;
    IsConfigNeeded: boolean;
    OwnField: ReadonlyArray<game_client.IOwnBattleFieldCell>;
    Enemy: IUserDto;
    EnemyField: ReadonlyArray<game_client.IEnemyBattleFieldCell>;
    IsOwnTurn: boolean;
    IsEnemyTurn: boolean;
}

export interface IJoinableMatchDto {
    Id: string;
    Creator: IUserDto;
}
// ...
export interface IPlayablesDto {
    CanCreateMatch: boolean;
    PendingMatchId: string;
    PlayingMatch: IMatchSnapshotDto;
    JoinableMatches: ReadonlyArray<IJoinableMatchDto>;
}

export interface IMatchEventDto {
    readonly MatchId: string;
}

/** When both players joined but someone didn't config his fleet yet **/
export interface IPendingMatchJoinedEventDto extends IMatchEventDto {
    readonly PendingMatchId: string;
}